requirements.txt
pandas==2.1.1
numpy==1.26.0

requests==2.31.0

python-dotenv==1.0.0

pyarrow==13.0.0  
fastparquet==2023.8.0  

matplotlib==3.8.0
seaborn==0.13.0

# Logging amélioré
loguru==0.7.2

# Tests
pytest==7.4.2
pytest-cov==4.1.0

# Linting et formatage
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Documentation
sphinx==7.2.6